requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
easyocr==1.7.1
pymongo==4.8.0
pymupdf==1.24.10
httpx==0.27.2
python-dotenv==1.0.1
